% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchMoSpaTA}
\alias{launchMoSpaTA}
\title{Launch MoSpaTA}
\usage{
launchMoSpaTA(...)
}
\value{
Shiny application.
}
\description{
Launch MoSpaTA
}
\details{
Launch MoSpaTA
}
\keyword{atlas}
\keyword{mouse}
\keyword{sc-RNASeq}
\keyword{spatialomics}
\keyword{testis}
